use simple string instead of object
because more complex the datatype the stronger the binding between the stuff.
By removing the dependency to the modulePathFormator things become more simple. E.g. unit test do not need to create that mock anymore.
The only downside is we do not have all information available anymore, but it's no big win to have the moduleid in the error message anyway because from the path you can guess find the module, or maybe broken settings in the database anyway.